Server/requirements.txt
Flask
pyjwt
bcrypt
pytest
psycopg2-binary
python-dotenv
pandas